+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v ?.??.? - ?? ??? ????
 - ShaderUtils prints shader program output information if using version 4.3+
+- Cylinder and Cone texture coords wrap outward textures
 
 ## v 5.17.0 - 06 Oct 2025
 - ShaderProgram can validate if program is ready to execute in current state